Update docker files for multiple runs
This change involved changing multiple src/main.py to specify the
parameters for multiple runs.

The compose fiel for the container running the code has changed to
reflect the change in the code. To run multiple instance of the same
container in the compose.yaml file you need to change the name of the
service, otherwise if there's a name conflict the older running
container will be shutdown.