Add file containing only extracted Answers for llama3b
I have created the file `llama3b_extracted_ans.jsonl` which contains
the extracted final answers from the llama3b answers, in this manner
we can repeat the LLM as judge only with this part of the answers. Here
we are using the LLM as a 'SentenceTransformer' to compute semantic
similarity.