Update Pipeline class + Fix bug on how I access messages in auto_reply_gen
The Pipeline class has been update to have two distinc methods consume
and to_list. Consume doesn't store the generated elements.

Fix bug on how i accessed the messages.
This is a bug coming from previous version. I was setting the result
payload as the messages, therfore I was overwriting the payload
(messages shoud have been an attribute of payload). This bug didn't
emerge in previous version, because the only function that was using the
output of auto_reply_gen was save_stream which accessed the payload as
if were the messages.