Add STARDataset class
The class handles the loading of QA (and other info of the STAR dataset)
and the association to a file with the STSGs. The stsg file should have
either a 'question_id', when the stsg is associate to a single question
or the set of values ('video_id', 'start', 'end') when is associated to
a video as STAR reference sub-videos of Charades, deliminting them by a
START and END time.