Add check for the existence + fill pd.NaN as None
When a request failes the resulting object does not contain the entry
`response`, but only an `error` entry. Therefore when loading the
response object to such request with pandas the response field is loaded
as pd.NaN, which is of type `float`. Therfore I convert the NaN as None
and when loading the response i first check the existence of the
`response` entry before accessing it.