Limit the fields in batch file for the 2nd step
When generating the batch file for the 2nd step limit the output fields
to just key and request, the only one required by the API. Before we
kept the response data of the previous responses as "documentation" but
it can lead to subtle bugs if one of the request fails. In this manner
we avoid this edge cases.